+++ usecases/type.go
@@ -6,5 +6,5 @@ import (
 
 type Wallet interface {
 	CreateOperation(paymentOperation *model.PaymentOperation) error
-	GetWalletBalanceByUUID(id string) (int, error)
+	GetWalletBalanceByUUID(id string) (int64, error)
 }
